Refactored molecule file created
